Move types into different mods, use differnt feature flag for testing
