+++ libs/types/src/lib.rs
@@ -20,7 +20,6 @@ pub mod consts;
 pub mod epoch;
 pub mod fee_keys;
 pub mod fixed_point;
-pub mod fulfillments;
 pub mod ids;
 pub mod investments;
 pub mod orders;
