+++ pallets/investments/src/mock.rs
@@ -17,8 +17,10 @@ pub use cfg_primitives::CFG as CURRENCY;
 use cfg_primitives::*;
 use cfg_traits::{OrderManager, PreConditions};
 use cfg_types::{
-	fixed_point::Rate, fulfillments::FulfillmentWithPrice, investments::InvestmentAccount,
-	orders::TotalOrder, tokens::CurrencyId,
+	fixed_point::Rate,
+	investments::InvestmentAccount,
+	orders::{FulfillmentWithPrice, TotalOrder},
+	tokens::CurrencyId,
 };
 use codec::{Decode, Encode};
 use frame_support::{
