+++ pallets/pool-system/src/benchmarking.rs
@@ -15,8 +15,8 @@
 use cfg_primitives::PoolEpochId;
 use cfg_traits::{InvestmentAccountant, InvestmentProperties, TrancheCurrency as _};
 use cfg_types::{
-	tokens::{CurrencyId, CustomMetadata},
-	tranches::{TrancheCurrency, TrancheIndex, TrancheInput, TrancheLoc, TrancheMetadata},
+	tokens::{CurrencyId, CustomMetadata, TrancheCurrency},
+	tranches::{TrancheIndex, TrancheInput, TrancheLoc, TrancheMetadata},
 };
 use codec::EncodeLike;
 use frame_benchmarking::{account, benchmarks, impl_benchmark_test_suite};
